refactor: migrate incident modals to helper module
